Refine cloud projection boundaries
