+++ packages/shared/src/index.ts
@@ -114,6 +114,7 @@ export * from './app-config.js'
 export * from './agent-validation.js'
 export * from './artifacts.js'
 export * from './capabilities.js'
+export * from './cloud-session-projection.js'
 export * from './custom-content.js'
 export * from './destructive-actions.js'
 export * from './events.js'
